refactor: restructure to pages/components and add @ alias
