+++ index.html
@@ -5,11 +5,11 @@
     <link rel="icon" type="image/svg+xml" href="/react.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Class components</title>
-    <link rel="stylesheet" href="src/style.css" />
+    <link rel="stylesheet" href="src/Style.css" />
   </head>
 
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/main.tsx"></script>
+    <script type="module" src="/src/Main.tsx"></script>
   </body>
 </html>
